use camera_info_manager to store/load camera parameters to/from file
The Azure Kinect publishes the factory calibrated intrinsic and extrinsic
camera parameters. The camera_info_manager provides the 'set_camera_info'
service to override these camera parameters within the ROS driver. This is
especially useful for a manual calibration of the RGB and IR cameras. After
custom camera parameters have been set, the ROS driver will publish those
in place of the factory calibration.
Note that this does not replace the factory calibration. The Kinect API in
the SDK will continue to use the factory calibration for rectification,
registration and 3D point cloud projection.